+++ Changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+### Version 1.19.3 - 1.4.4
+- fix forge creative tab crash
+
 ### Version 1.19.3 - 1.4.3
 - port to 1.19.3
 - ToolLeveling Table is now in the creative tab of functional blocks